+++ src/main/java/kitchenpos/products/domain/JpaProductRepository.java
@@ -1,6 +1,7 @@
 package kitchenpos.products.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import kitchenpos.products.tobe.domain.Product;
 
 import java.util.UUID;
