+++ src/main/java/kitchenpos/products/domain/ProductRepository.java
@@ -3,6 +3,7 @@ package kitchenpos.products.domain;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.products.tobe.domain.Product;
 
 public interface ProductRepository {
     Product save(Product product);
